+++ cmd/project.go
@@ -8,8 +8,8 @@ import (
 
 var projectCmd = &cobra.Command{
 	Use:   "project [command]",
-	Short: "manage projects",
-	Long:  "Project management for Runpod projects",
+	Short: "Manage RunPod projects",
+	Long:  "Develop and deploy projects entirely on RunPod's infrastructure",
 }
 
 func init() {